algunos arreglos

+++ app/src/main/java/com/easypick/easypick/retroFit/Gateway.kt
@@ -13,7 +13,7 @@ import java.util.*
 interface Gateway {
     @GET("api.php")
     fun getProductoByCategoryId(@Query("store") store: Long,
-                                @Query("categoria") categoria: Long): Call<List<Producto>>
+                                @Query("categoria") categoria: Int): Call<List<Producto>>
 
     @GET("/stores/{id}/categories")
     fun getCategoryByStoreId(@Path("id") id:Long): Call<List<Category>>
